requirements.txt
pandas==2.1.4
pymongo==4.6.1
python-dotenv==1.0.0
colorama==0.4.6
